+++ server/src/services/script/ComponentPrivateService.ts
@@ -2,7 +2,7 @@ import { BaseService } from "../Common";
 import { SvelteDocument } from "../../SvelteDocument";
 import { CompletionItem } from "vscode-languageserver";
 import { ScopeContext } from "../../interfaces";
-import { DefaultComponentMethods, DefaultRefCompletionItem, DefaultComponentGetMethodCompletionItem } from "../../svelteLanguage";
+import { DefaultComponentMethods, DefaultRefCompletionItem, DefaultComponentGetMethodCompletionItem, DefaultScriptRefsCompletionItem } from "../../svelteLanguage";
 import { cloneCompletionItem } from "../Utils";
 
 export class ComponentPrivateService extends BaseService {
@@ -12,7 +12,7 @@ export class ComponentPrivateService extends BaseService {
             return  [
                 ...DefaultComponentMethods,
                 DefaultComponentGetMethodCompletionItem,
-                DefaultRefCompletionItem,
+                DefaultScriptRefsCompletionItem,
                 ...document.metadata.methods,
                 ...document.metadata.refs
                     .map(cloneCompletionItem)
@@ -33,14 +33,11 @@ export class ComponentPrivateService extends BaseService {
                 .map(cloneCompletionItem)
                 .map(item => {
                     item.detail = `[Svelte] Reference to element/component`;
-                    item.filterText = `refs.${item.label}`;
-                    item.sortText = `refs.${item.label}`;
-                    item.insertText = `refs.${item.label}`;
-
                     return item;
                 })
             ];
         }
+
         return null;
     }
 }