+++ server/src/services/style/RefsStyleService.ts
@@ -6,24 +6,24 @@ import { DefaultRefCompletionItem } from "../../svelteLanguage";
 
 export class RefsStyleService extends BaseService {
     public getCompletitionItems(document: SvelteDocument): Array<CompletionItem> {
-        let result = [];
-        
-        result.push(...[
+        let result = [
             DefaultRefCompletionItem
-        ]);
-
-        result.push(
-            ...document.metadata.refs
-            .map(cloneCompletionItem)
-            .map(item => {
-                item.filterText = `ref:${item.label}`;
-                item.sortText = `ref:${item.label}`;
-                item.insertText = `ref:${item.label}`;
-                item.detail = `[Svelte] ref:${item.label}`;
-                item.commitCharacters = [' '];
-                return item;
-            })
-        );
+        ];
+        
+        if (document.metadata) {
+            result.push(
+                ...document.metadata.refs
+                .map(cloneCompletionItem)
+                .map(item => {
+                    item.filterText = `ref:${item.label}`;
+                    item.sortText = `ref:${item.label}`;
+                    item.insertText = `ref:${item.label}`;
+                    item.detail = `[Svelte] ref:${item.label}`;
+                    item.commitCharacters = [' '];
+                    return item;
+                })
+            );
+        }
 
         return result;
     }