+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 - [Fixed] Fix issue #2 with `:catch` and `:then` svelte syntax auto-insertion
 - [Fixed] Fix issue #3 with `ref:`, `transition:`, `in:`, `out:`, `use:` auto-insertion when user press <kbd>:</kbd>
+- [Fixed] Improve auto-completion for `this.`
+- [Added] Add auto completion for single property getter `this.get().<>`
 - [Added] Auto-completion for data and computed properties and helper methods in `{#if _}`, `{:elseif _}`, `{#each _}`, `{#await _}` statements
 
 ## [0.2.0]